pyproject.toml
[tool.poetry]
name = "face-detection-api"
version = "0.1.0"
description = ""
authors = ["Glendell Bringino dellbringino@gmail.com"]
readme = "README.md"
packages = [{include = "face_detection_api"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.95.2"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
prisma = "^0.8.2"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
supabase = "^1.0.3"
twilio = "^8.2.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
